backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pydantic
pydantic-settings
langchain
langchain-openai
google-generativeai
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-multipart
click
emails
